// frontend/src/app/swagger/model/boardCreateDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstanceCreateDto } from './instanceCreateDto';


export interface BoardCreateDto { 
    name?: string | null;
    instances?: Array<InstanceCreateDto> | null;
}