// frontend/src/app/swagger/model/instanceCreateDto.ts
/**
 * S2Retro.Presentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InstanceColumnCreateDto } from './instanceColumnCreateDto';
import { InstancePhase } from './instancePhase';
import { InstanceRowCreateDto } from './instanceRowCreateDto';


export interface InstanceCreateDto { 
    name?: string | null;
    entryPhaseStart?: string;
    entryPhaseEnd?: string | null;
    phase?: InstancePhase;
    boardId?: string;
    columns?: Array<InstanceColumnCreateDto> | null;
    rows?: Array<InstanceRowCreateDto> | null;
}
export namespace InstanceCreateDto {
}